refactor: 미사용 코드 제거 및 내피드 총 갯수 표기
