+++ domain/feed/src/main/java/com/into/websoso/feed/GetMyFeedsUseCase.kt
@@ -1,17 +1,26 @@
 package com.into.websoso.feed
 
+import com.into.websoso.data.feed.repository.UpdatedFeedRepository
 import com.into.websoso.feed.mapper.toDomain
+import com.into.websoso.feed.mapper.toFeedEntity
+import com.into.websoso.feed.model.Feed
 import com.into.websoso.feed.model.Feeds
 import com.into.websoso.user.UserRepository
 import com.into.websoso.user.model.MyProfileEntity
 import com.into.websoso.user.model.UserFeedsEntity
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
 class GetMyFeedsUseCase
     @Inject
     constructor(
         private val userRepository: UserRepository,
+        private val feedRepository: UpdatedFeedRepository,
     ) {
+        val myFeedsFlow: Flow<List<Feed>> = feedRepository.myFeeds
+            .map { list -> list.map { it.toDomain() } }
+
         private var myProfile: MyProfileEntity? = null
         private var myId: Long? = null
 
@@ -27,7 +36,7 @@ class GetMyFeedsUseCase
             val profile = myProfile ?: userRepository.fetchMyProfile().also { myProfile = it }
             val myId = myId ?: userRepository.fetchUserInfo().userId.also { myId = it }
 
-            val myFeeds: UserFeedsEntity = userRepository.fetchMyActivities(
+            val myFeedsEntity: UserFeedsEntity = userRepository.fetchMyActivities(
                 lastFeedId = lastFeedId,
                 size = if (isFeedRefreshed) INITIAL_REQUEST_SIZE else ADDITIONAL_REQUEST_SIZE,
                 genres = genres?.toTypedArray(),
@@ -36,10 +45,20 @@ class GetMyFeedsUseCase
                 sortCriteria = sortCriteria,
             )
 
+            val convertedFeeds = myFeedsEntity.feeds.map { userFeed ->
+                userFeed.toFeedEntity(userProfile = profile, userId = myId)
+            }
+
+            feedRepository.updateMyFeedsCache(
+                feeds = convertedFeeds,
+                isRefreshed = isFeedRefreshed,
+            )
+
             return Feeds(
                 category = "내 활동",
-                isLoadable = myFeeds.isLoadable,
-                feeds = myFeeds.feeds.map { it.toDomain(myProfile = profile, id = myId) },
+                isLoadable = myFeedsEntity.isLoadable,
+                totalCount = myFeedsEntity.feedsCount,
+                feeds = emptyList(),
             )
         }
