+++ feature/feed/src/main/java/com/into/websoso/feature/feed/FeedScreen.kt
@@ -91,7 +91,7 @@ internal fun FeedScreen(
                 when (uiState.selectedTab) {
                     FeedTab.MY_FEED -> {
                         FeedFilterChip(
-                            label = "${uiState.myFeedData.feeds.size}개의 기록",
+                            label = "${uiState.myFeedData.totalCount}개의 기록",
                             isSelected = true,
                             rightIcon = {
                                 Icon(
@@ -222,7 +222,8 @@ private fun FeedTabRow(
                         .tabIndicatorOffset(
                             selectedTabIndex = selectedTab.ordinal,
                             matchContentSize = false,
-                        ).padding(horizontal = 8.dp),
+                        )
+                        .padding(horizontal = 8.dp),
                     height = 2.dp,
                     color = Black,
                 )
